File conftest.py added
