+++ tests/conftest.py
@@ -1,0 +1,41 @@
+import pytest
+
+
+@pytest.fixture
+def number_account() -> str:
+    return "**4305"
+
+
+@pytest.fixture
+def account() -> str:
+    return "Счет **5560"
+
+
+@pytest.fixture
+def card() -> str:
+    return "Visa Platinum 8990 92** **** 5229"
+
+
+@pytest.fixture
+def date() -> str:
+    return "11.03.2024"
+
+
+@pytest.fixture
+def my_lists() -> list[dict]:
+    return [
+        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
+        {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
+        {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
+        {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
+    ]
+
+
+@pytest.fixture
+def dates() -> list[dict]:
+    return [
+        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
+        {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
+        {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
+        {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
+    ]
