+++ tests/test_processing.py
@@ -3,16 +3,6 @@ import pytest
 from src.processing import filter_by_state, sort_by_date
 
 
-@pytest.fixture
-def my_lists() -> list[dict]:
-    return [
-        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-        {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
-        {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
-        {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
-    ]
-
-
 @pytest.mark.parametrize(
     "state, expected",
     [
@@ -39,16 +29,6 @@ def test_filter_by_state(my_lists: list[dict], state: str, expected: list[dict])
 @pytest.mark.parametrize("list_dicts", [[{"id": 594226727, "date": "2018-09-12T21:27:25.241689"}], [{}]])
 def test_filter_by_state_invalid(list_dicts: list[dict]) -> None:
     assert filter_by_state(list_dicts) == []
-
-
-@pytest.fixture
-def dates() -> list[dict]:
-    return [
-        {"id": 41428829, "state": "EXECUTED", "date": "2019-07-03T18:35:29.512364"},
-        {"id": 939719570, "state": "EXECUTED", "date": "2018-06-30T02:08:58.425572"},
-        {"id": 594226727, "state": "CANCELED", "date": "2018-09-12T21:27:25.241689"},
-        {"id": 615064591, "state": "CANCELED", "date": "2018-10-14T08:21:33.419441"},
-    ]
 
 
 @pytest.mark.parametrize(
